Adds logic to delete product image when the product is deleted

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const fs = require('fs');
+const fileHelper = require('../util/file');
 
 const { validationResult } =require('express-validator/check');
+
 
 const Product = require('../models/product');
 
@@ -145,6 +147,7 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       product.description = updatedDesc;
@@ -165,8 +168,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  // Product.findByIdAndRemove(prodId)
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+  .then(product => {
+    if(!product) {
+      return next(new Error('Product not found.'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  })
   .then(() => {
     console.log('DESTROYED PRODUCT!');
     res.redirect('/admin/products');
